+++ sonarts-core/src/utils/sonar-analysis.ts
